requirements.txt
absl-py==0.8.1
astor==0.8.0
backcall==0.1.0
bleach==1.5.0
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cycler==0.10.0
Cython==0.29.13
dataclasses==0.7
decorator==4.4.0
fastapi==0.43.0
gast==0.2.2
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
graphsurgeon==0.4.1
graphviz==0.8.4
grpcio==1.25.0
h11==0.8.1
h5py==2.9.0
html5lib==0.9999999
httptools==0.0.13
idna==2.8
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
joblib==0.14.1
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.4.1
Markdown==3.1.1
matplotlib==3.1.2
munch==2.5.0
mxnet-cu100==1.4.0.post0
numpy==1.16.0
oauthlib==3.1.0
object-detection==0.1
opencv-python==4.1.1.26
opt-einsum==3.0.1
pandas==0.25.3
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
pretrainedmodels==0.7.4
prompt-toolkit==2.0.9
protobuf==3.9.1
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pydantic==0.32.2
Pygments==2.3.1
pyparsing==2.4.5
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22
scipy==1.3.3
six==1.13.0
slim==0.1
starlette==0.12.9
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
torch==1.2.0
torchvision==0.4.0
tqdm==4.37.0
traitlets==4.3.2
typing-extensions==3.7.4.1
urllib3==1.25.7
uvicorn==0.10.8
uvloop==0.14.0
wcwidth==0.1.7
websockets==8.1
Werkzeug==0.16.0
wrapt==1.11.1
